Lab 10 finished by U1710010
Tagoev Ozod 
U1710010
CSE -17-1
Section 001

+++ app/Http/Controllers/PostController.php
@@ -1,12 +1,11 @@
 <?php
-
 namespace App\Http\Controllers;
-
+use Auth;
+use Gate;
 use App\Like;
 use App\Post;
 use App\Tag;
 use Illuminate\Http\Request;
-
 class PostController extends Controller
 {
     public function getIndex()
@@ -14,19 +13,16 @@ class PostController extends Controller
         $posts = Post::orderBy('created_at', 'desc')->paginate(2);
         return view('blog.index', ['posts' => $posts]);
     }
-
     public function getAdminIndex()
     {
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
-
     public function getPost($id)
     {
         $post = Post::where('id', $id)->with('likes')->first();
         return view('blog.post', ['post' => $post]);
     }
-
     public function getLikePost($id)
     {
         $post = Post::where('id', $id)->first();
@@ -34,36 +30,32 @@ class PostController extends Controller
         $post->likes()->save($like);
         return redirect()->back();
     }
-
     public function getAdminCreate()
     {
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
-
     public function getAdminEdit($id)
     {
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
     }
-
     public function postAdminCreate(Request $request)
     {
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        $user = Auth::user();
         $post = new Post([
             'title' => $request->input('title'),
             'content' => $request->input('content')
         ]);
-        $post->save();
+        $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
-
         return redirect()->route('admin.index')->with('info', 'Post created, Title is: ' . $request->input('title'));
     }
-
     public function postAdminUpdate(Request $request)
     {
         $this->validate($request, [
@@ -71,6 +63,9 @@ class PostController extends Controller
             'content' => 'required|min:10'
         ]);
         $post = Post::find($request->input('id'));
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -79,10 +74,12 @@ class PostController extends Controller
         $post->tags()->sync($request->input('tags') === null ? [] : $request->input('tags'));
         return redirect()->route('admin.index')->with('info', 'Post edited, new Title is: ' . $request->input('title'));
     }
-
     public function getAdminDelete($id)
     {
         $post = Post::find($id);
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
